pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dpt2"
version = "0.1.0"
description = "Multiplex dPCR primer/probe design tool"
authors = [{ name = "Marie" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "primer3-py>=2.0.0",
    "biopython>=1.80",
    "pandas>=1.3.0",
    "pyyaml>=6.0"
]

[project.scripts]
design-primers = "mpx_dpcr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mpx_dpcr"]

